Norm almost done, float become double, but did not correct weird line hqlf screen

+++ srcs/player.h
@@ -6,7 +6,7 @@
 /*   By: lchapren <lchapren@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/19 13:36:00 by lchapren          #+#    #+#             */
-/*   Updated: 2020/08/12 10:53:09 by lchapren         ###   ########.fr       */
+/*   Updated: 2020/08/18 15:02:41 by lchapren         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,16 +15,16 @@
 
 typedef struct	s_player
 {
-	float	pos_x;
-	float	pos_y;
-	float	dir_x;
-	float	dir_y;
-	float	plane_x;
-	float	plane_y;
+	double	pos_x;
+	double	pos_y;
+	double	dir_x;
+	double	dir_y;
+	double	plane_x;
+	double	plane_y;
 
-	float	speed;
-	float	rotation_angle;
-	float	height;
+	double	speed;
+	double	rotation_angle;
+	double	height;
 
 	int		foward;
 	int		left;
